Refactor: extract sliding window from SearchWorkerResult.kt
The code is still a bit buggy and tests are not yet
adapted to new file disposition
