+++ app/src/main/java/com/arygm/quickfix/ui/search/SearchWorkerResult.kt
@@ -1,49 +1,33 @@
+// Updated SearchWorkerResult.kt
+
 package com.arygm.quickfix.ui.search
 
+import QuickFixSlidingWindowWorker
 import android.location.Location
 import android.util.Log
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.verticalScroll
+import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.Bookmark
 import androidx.compose.material.icons.filled.CalendarMonth
 import androidx.compose.material.icons.filled.KeyboardArrowDown
 import androidx.compose.material.icons.filled.Search
-import androidx.compose.material.icons.filled.Star
 import androidx.compose.material.icons.filled.Tune
 import androidx.compose.material.icons.filled.WorkspacePremium
-import androidx.compose.material.icons.outlined.BookmarkBorder
-import androidx.compose.material.icons.outlined.StarOutline
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.IconButtonDefaults
@@ -53,7 +37,6 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -61,24 +44,19 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.arygm.quickfix.MainActivity
 import com.arygm.quickfix.R
-import com.arygm.quickfix.model.account.Account
 import com.arygm.quickfix.model.account.AccountViewModel
 import com.arygm.quickfix.model.search.SearchViewModel
 import com.arygm.quickfix.ui.elements.QuickFixAvailabilityBottomSheet
 import com.arygm.quickfix.ui.elements.QuickFixButton
-import com.arygm.quickfix.ui.elements.QuickFixSlidingWindow
 import com.arygm.quickfix.ui.navigation.NavigationActions
 import com.arygm.quickfix.ui.theme.poppinsTypography
 import com.arygm.quickfix.utils.LocationHelper
@@ -98,6 +76,8 @@ fun SearchWorkerResult(
     accountViewModel: AccountViewModel
 ) {
   var showAvailabilityBottomSheet by remember { mutableStateOf(false) }
+  var isWindowVisible by remember { mutableStateOf(false) }
+  var saved by remember { mutableStateOf(false) }
 
   val listOfButtons =
       listOf(
@@ -127,85 +107,33 @@ fun SearchWorkerResult(
           ),
       )
 
-  // ==========================================================================//
-  // ============ TODO: REMOVE NO-DATA WHEN BACKEND IS IMPLEMENTED ============//
-  // ==========================================================================//
-
-  val bannerImage = R.drawable.moroccan_flag
-  val profilePicture = R.drawable.placeholder_worker
-  val workerName = "Moha Abbes"
-  val workerCategory = "Exterior Painter"
-  val description =
-      "According to all known laws of aviation, there is no way a bee should be able " +
-          "to fly. Its wings are too small to get its fat little body off the ground. The bee, of " +
-          "course, flies anyway because bees don't care what humans think is impossible. Yellow, " +
-          "black. Yellow, black. Yellow, black. Yellow, black. Ooh, black and yellow! Let's shake " +
-          "it up a little. Barry! Breakfast is ready! Coming! Hang on a second. Hello? - Barry? - Adam? " +
-          "- Can you believe this is happening? - I can't. I'll pick you up. Looking sharp. Use the stairs. " +
-          "Your father paid good money for those. Sorry. I'm excited. Here's the graduate. We're " +
-          "very proud of you, son. A perfect report card, all B's. Very proud. Ma! I got a thing " +
-          "going here. - You got lint on your fuzz. - Ow! That's me! - Wave to us! We'll be in row " +
-          "118,000. - Bye! Barry, I told you, stop flying in the house! - Hey, Adam. - Hey, Barry. " +
-          "- Is that fuzz gel? - A little. Special day, graduation. Never thought I'd make it. " +
-          "Three days grade school, three days high school. Those were awkward. Three days college. " +
-          "I'm glad I took a day and hitchhiked around the hive. You did come back different. " +
-          "- Hi, Barry. - Artie, growing a mustache? Looks good. - Hear about Frankie? - Yeah. " +
-          "- You going to the funeral? - No, I'm not going. Everybody knows, sting someone, you die. " +
-          "Don't waste it on a squirrel. Such a hothead. I guess he could have just gotten out of the way. " +
-          "I love this incorporating an amusement park into our day. That's why we don't need vacations. " +
-          "Boy, quite a bit of pomp... under the circumstances."
-  val workerAddress = "Ecublens, VD"
-  val workerRating = 3.8
-  val includedServices =
-      listOf(
-          "Initial Consultation",
-          "Basic Surface Preparation",
-          "Priming of Surfaces",
-          "High-Quality Paint Application",
-          "Two Coats of Paint",
-          "Professional Cleanup")
-  val addonServices =
-      listOf(
-          "Detailed Color Consultation",
-          "Premium paint Upgrade",
-          "Extensive Surface Preparation",
-          "Extra Coats for added Durability",
-          "Power Washing and Deep Cleaning")
-  val rate = 40
-  val tags =
-      listOf(
-          "Exterior Painting",
-          "Interior Painting",
-          "Cabinet Painting",
-          "Licensed & Insured",
-          "Local Worker")
-  val reviews =
-      listOf(
-          "Overall, the work was shit; it’s fair to say that since he just painted the whole mural with " +
-              "his shit. I don’t know what to do; I’ll probably sue him.",
-          "Moha was very punctual and did a great job painting our living room. Highly recommended!",
-          "Moha was very professional and did a fantastic job painting our house",
-          "I wanna marry that man,")
-
-  // ==========================================================================//
-  // ==========================================================================//
-  // ==========================================================================//
-
-  var isWindowVisible by remember { mutableStateOf(false) }
-  var saved by remember { mutableStateOf(false) }
   val searchQuery by searchViewModel.searchQuery.collectAsState()
   val workerProfiles by searchViewModel.workerProfiles.collectAsState()
   var currentLocation by remember { mutableStateOf<Location?>(null) }
   var filteredWorkerProfiles by remember { mutableStateOf(workerProfiles) }
 
   val locationHelper: LocationHelper = LocationHelper(LocalContext.current, MainActivity())
+  val listState = rememberLazyListState()
+
+  // Variables for Sliding Window
+  var bannerImage by remember { mutableStateOf(R.drawable.moroccan_flag) }
+  var profilePicture by remember { mutableStateOf(R.drawable.placeholder_worker) }
+  var initialSaved by remember { mutableStateOf(false) }
+  var workerCategory by remember { mutableStateOf("Exterior Painter") }
+  var workerAddress by remember { mutableStateOf("Ecublens, VD") }
+  var description by remember { mutableStateOf("Worker description goes here.") }
+  var includedServices by remember { mutableStateOf(listOf<String>()) }
+  var addonServices by remember { mutableStateOf(listOf<String>()) }
+  var workerRating by remember { mutableStateOf(4.5) }
+  var tags by remember { mutableStateOf(listOf<String>()) }
+  var reviews by remember { mutableStateOf(listOf<String>()) }
 
   // Wrap everything in a Box to allow overlay
   BoxWithConstraints(modifier = Modifier.fillMaxSize()) {
     val screenHeight = maxHeight
     val screenWidth = maxWidth
     Log.d("Screen Dimensions", "Height: $screenHeight, Width: $screenWidth")
-    // Scaffold containing the main UI elements
+
     Scaffold(
         topBar = {
           CenterAlignedTopAppBar(
@@ -267,22 +195,22 @@ fun SearchWorkerResult(
                     verticalAlignment = Alignment.CenterVertically,
                 ) {
                   items(1) {
-                    Box(modifier = Modifier.height(screenHeight * 0.05f)) {
-                      IconButton(
-                          onClick = { /* Goes to all filter screen */},
-                          modifier = Modifier.padding(bottom = screenHeight * 0.01f),
-                          content = {
-                            Icon(
-                                imageVector = Icons.Default.Tune,
-                                contentDescription = "Filter",
-                                tint = colorScheme.onBackground,
-                            )
-                          },
-                          colors =
-                              IconButtonDefaults.iconButtonColors()
-                                  .copy(containerColor = colorScheme.surface),
-                      )
-                    }
+                    IconButton(
+                        onClick = { /* Goes to all filter screen */},
+                        modifier =
+                            Modifier.height(screenHeight * 0.05f)
+                                .padding(bottom = screenHeight * 0.01f),
+                        content = {
+                          Icon(
+                              imageVector = Icons.Default.Tune,
+                              contentDescription = "Filter",
+                              tint = colorScheme.onBackground,
+                          )
+                        },
+                        colors =
+                            IconButtonDefaults.iconButtonColors()
+                                .copy(containerColor = colorScheme.surface),
+                    )
                     Spacer(modifier = Modifier.width(10.dp))
                   }
 
@@ -304,57 +232,39 @@ fun SearchWorkerResult(
                   }
                 }
 
-                LazyColumn(modifier = Modifier.fillMaxWidth().testTag("worker_profiles_list")) {
-                  items(filteredWorkerProfiles.size) { index ->
-                    val profile = filteredWorkerProfiles[index]
-                    var account by remember { mutableStateOf<Account?>(null) }
-                    var distance by remember { mutableStateOf<Int?>(null) }
+                // Use ProfileResults here
+                ProfileResults(
+                    profiles = filteredWorkerProfiles,
+                    listState = listState,
+                    searchViewModel = searchViewModel,
+                    accountViewModel = accountViewModel,
+                    widthRatio = 1f, // Adjust as needed
+                    heightRatio = 1f, // Adjust as needed
+                    onBookClick = { selectedProfile ->
+                      // Handle profile click
+                      // Update variables for Sliding Window
+                      bannerImage = R.drawable.moroccan_flag // Replace with actual data
+                      profilePicture = R.drawable.placeholder_worker // Replace with actual data
+                      initialSaved = false // Replace with actual data
+                      workerCategory = selectedProfile.fieldOfWork
+                      workerAddress = selectedProfile.location?.name ?: "Unknown"
+                      description = selectedProfile.description
 
-                    locationHelper.getCurrentLocation { location ->
-                      currentLocation = location
-                      location?.let {
-                        distance =
-                            profile.location
-                                ?.let { workerLocation ->
-                                  searchViewModel.calculateDistance(
-                                      workerLocation.latitude,
-                                      workerLocation.longitude,
-                                      it.latitude,
-                                      it.longitude)
-                                }
-                                ?.toInt()
-                      }
-                    }
+                      // Extract included services
+                      includedServices = selectedProfile.includedServices.map { it.name }
 
-                    LaunchedEffect(profile.uid) {
-                      accountViewModel.fetchUserAccount(profile.uid) { fetchedAccount: Account? ->
-                        account = fetchedAccount
-                      }
-                    }
+                      // Extract addon services
+                      addonServices = selectedProfile.addOnServices.map { it.name }
 
-                    account?.let { acc ->
-                      val locationName =
-                          if (profile.location?.name.isNullOrEmpty()) "Unknown"
-                          else profile.location?.name
+                      workerRating = selectedProfile.rating
+                      tags = selectedProfile.tags
 
-                      locationName?.let {
-                        SearchWorkerProfileResult(
-                            modifier = Modifier.testTag("worker_profile_result$index"),
-                            profileImage = R.drawable.placeholder_worker,
-                            name = "${acc.firstName} ${acc.lastName}",
-                            category = profile.fieldOfWork,
-                            rating = profile.rating,
-                            reviewCount = profile.reviews.size,
-                            location = it,
-                            price = profile.price.toString(),
-                            onBookClick = { isWindowVisible = true },
-                            distance = distance,
-                        )
-                      }
-                    }
-                    Spacer(modifier = Modifier.height(screenHeight * 0.004f))
-                  }
-                }
+                      // Extract reviews
+                      reviews =
+                          selectedProfile.reviews.map { it.review } // Use the correct property
+
+                      isWindowVisible = true
+                    })
               }
         }
 
@@ -367,318 +277,23 @@ fun SearchWorkerResult(
               searchViewModel.filterWorkersByAvailability(workerProfiles, days, hour, minute)
         }
 
-      QuickFixSlidingWindow(isVisible = isWindowVisible, onDismiss = { isWindowVisible = false }) {
-        // Content of the sliding window
-        Column(
-            modifier =
-                Modifier.clip(RoundedCornerShape(topStart = 25f, bottomStart = 25f))
-                    .fillMaxWidth()
-                    .background(colorScheme.background)
-                    .testTag("sliding_window_content")) {
-
-              // Top Bar
-              Box(
-                  modifier =
-                      Modifier.fillMaxWidth()
-                          .height(
-                              screenHeight *
-                                  0.23f) // Adjusted height to accommodate profile picture overlap
-                          .testTag("sliding_window_top_bar")) {
-                    // Banner Image
-                    Image(
-                        painter = painterResource(id = bannerImage),
-                        contentDescription = "Banner",
-                        modifier =
-                            Modifier.fillMaxWidth()
-                                .height(screenHeight * 0.2f)
-                                .testTag("sliding_window_banner_image"),
-                        contentScale = ContentScale.Crop)
-
-                    QuickFixButton(
-                        buttonText = if (saved) "saved" else "save",
-                        onClickAction = { saved = !saved },
-                        buttonColor = colorScheme.surface,
-                        textColor = colorScheme.onBackground,
-                        textStyle = MaterialTheme.typography.labelMedium,
-                        contentPadding = PaddingValues(horizontal = screenWidth * 0.01f),
-                        modifier =
-                            Modifier.align(Alignment.BottomEnd)
-                                .width(screenWidth * 0.25f)
-                                .offset(x = -(screenWidth * 0.04f))
-                                .testTag(
-                                    "sliding_window_save_button"), // Negative offset to position
-                        // correctly,
-                        leadingIcon =
-                            if (saved) Icons.Filled.Bookmark else Icons.Outlined.BookmarkBorder)
-
-                    // Profile picture overlapping the banner image
-                    Image(
-                        painter = painterResource(id = profilePicture),
-                        contentDescription = "Profile Picture",
-                        modifier =
-                            Modifier.size(screenHeight * 0.1f)
-                                .align(Alignment.BottomStart)
-                                .offset(x = screenWidth * 0.04f)
-                                .clip(CircleShape)
-                                .testTag("sliding_window_profile_picture"),
-                        // Negative offset to position correctly
-                        contentScale = ContentScale.Crop)
-                  }
-
-              // Worker Field and Address under the profile picture
-              Column(
-                  modifier =
-                      Modifier.fillMaxWidth()
-                          .padding(horizontal = screenWidth * 0.04f)
-                          .testTag("sliding_window_worker_additional_info")) {
-                    Text(
-                        text = workerCategory,
-                        style = MaterialTheme.typography.headlineLarge,
-                        color = colorScheme.onBackground,
-                        modifier = Modifier.testTag("sliding_window_worker_category"))
-                    Text(
-                        text = workerAddress,
-                        style = MaterialTheme.typography.headlineSmall,
-                        color = colorScheme.onBackground,
-                        modifier = Modifier.testTag("sliding_window_worker_address"))
-                  }
-
-              // Main content should be scrollable
-              Column(
-                  modifier =
-                      Modifier.fillMaxWidth()
-                          .verticalScroll(rememberScrollState())
-                          .background(colorScheme.surface)
-                          .testTag("sliding_window_scrollable_content")) {
-                    Spacer(modifier = Modifier.height(screenHeight * 0.02f))
-
-                    // Description with "Show more" functionality
-                    var showFullDescription by remember { mutableStateOf(false) }
-                    val descriptionText =
-                        if (showFullDescription || description.length <= 100) {
-                          description
-                        } else {
-                          description.take(100) + "..."
-                        }
-
-                    Text(
-                        text = descriptionText,
-                        style = MaterialTheme.typography.bodySmall,
-                        color = colorScheme.onSurface,
-                        modifier =
-                            Modifier.padding(horizontal = screenWidth * 0.04f)
-                                .testTag("sliding_window_description"))
-
-                    if (description.length > 100) {
-                      Text(
-                          text = if (showFullDescription) "Show less" else "Show more",
-                          style =
-                              MaterialTheme.typography.bodySmall.copy(color = colorScheme.primary),
-                          modifier =
-                              Modifier.padding(horizontal = screenWidth * 0.04f)
-                                  .clickable { showFullDescription = !showFullDescription }
-                                  .testTag("sliding_window_description_show_more_button"))
-                    }
-
-                    // Delimiter between description and services
-                    Spacer(modifier = Modifier.height(screenHeight * 0.02f))
-
-                    HorizontalDivider(
-                        modifier =
-                            Modifier.padding(horizontal = screenWidth * 0.04f)
-                                .testTag("sliding_window_horizontal_divider_1"),
-                        thickness = 1.dp,
-                        color = colorScheme.onSurface.copy(alpha = 0.2f))
-                    Spacer(modifier = Modifier.height(screenHeight * 0.02f))
-
-                    // Services Section
-                    Row(
-                        modifier =
-                            Modifier.fillMaxWidth()
-                                .padding(horizontal = screenWidth * 0.04f)
-                                .testTag("sliding_window_services_row")) {
-                          // Included Services
-                          Column(
-                              modifier =
-                                  Modifier.weight(1f)
-                                      .testTag("sliding_window_included_services_column")) {
-                                Text(
-                                    text = "Included Services",
-                                    style = MaterialTheme.typography.headlineMedium,
-                                    color = colorScheme.onBackground)
-                                Spacer(modifier = Modifier.height(screenHeight * 0.01f))
-                                includedServices.forEach { service ->
-                                  Text(
-                                      text = "• $service",
-                                      style = MaterialTheme.typography.bodySmall,
-                                      color = colorScheme.onSurface,
-                                      modifier = Modifier.padding(bottom = screenHeight * 0.005f))
-                                }
-                              }
-
-                          Spacer(modifier = Modifier.width(screenWidth * 0.02f))
-
-                          // Add-On Services
-                          Column(
-                              modifier =
-                                  Modifier.weight(1f)
-                                      .testTag("sliding_window_addon_services_column")) {
-                                Text(
-                                    text = "Add-On Services",
-                                    style = MaterialTheme.typography.headlineMedium,
-                                    color = colorScheme.primary)
-                                Spacer(modifier = Modifier.height(screenHeight * 0.01f))
-                                addonServices.forEach { service ->
-                                  Text(
-                                      text = "• $service",
-                                      style = MaterialTheme.typography.bodySmall,
-                                      color = colorScheme.primary,
-                                      modifier = Modifier.padding(bottom = screenHeight * 0.005f))
-                                }
-                              }
-                        }
-
-                    Spacer(modifier = Modifier.height(screenHeight * 0.03f))
-
-                    // Continue Button with Rate/HR
-                    QuickFixButton(
-                        buttonText = "Continue",
-                        onClickAction = { /* Handle continue */},
-                        buttonColor = colorScheme.primary,
-                        textColor = colorScheme.onPrimary,
-                        textStyle = MaterialTheme.typography.labelMedium,
-                        modifier =
-                            Modifier.fillMaxWidth()
-                                .padding(horizontal = screenWidth * 0.04f)
-                                .testTag("sliding_window_continue_button"))
-
-                    Spacer(modifier = Modifier.height(screenHeight * 0.02f))
-
-                    HorizontalDivider(
-                        modifier =
-                            Modifier.padding(horizontal = screenWidth * 0.04f)
-                                .testTag("sliding_window_horizontal_divider_2"),
-                        thickness = 1.dp,
-                        color = colorScheme.onSurface.copy(alpha = 0.2f),
-                    )
-                    Spacer(modifier = Modifier.height(screenHeight * 0.02f))
-
-                    // Tags Section
-                    Text(
-                        text = "Tags",
-                        style = MaterialTheme.typography.headlineMedium,
-                        color = colorScheme.onBackground,
-                        modifier = Modifier.padding(horizontal = screenWidth * 0.04f))
-                    Spacer(modifier = Modifier.height(screenHeight * 0.01f))
-
-                    // Display tags using FlowRow for wrapping
-                    FlowRow(
-                        horizontalArrangement = Arrangement.spacedBy(screenWidth * 0.02f),
-                        verticalArrangement = Arrangement.spacedBy(screenHeight * 0.01f),
-                        modifier =
-                            Modifier.fillMaxWidth()
-                                .padding(horizontal = screenWidth * 0.04f)
-                                .testTag("sliding_window_tags_flow_row"),
-                    ) {
-                      tags.forEach { tag ->
-                        Text(
-                            text = tag,
-                            color = colorScheme.primary,
-                            style = MaterialTheme.typography.bodySmall,
-                            modifier =
-                                Modifier.border(
-                                        width = 1.dp,
-                                        color = colorScheme.primary,
-                                        shape = MaterialTheme.shapes.small)
-                                    .padding(
-                                        horizontal = screenWidth * 0.02f,
-                                        vertical = screenHeight * 0.005f))
-                      }
-                    }
-
-                    Spacer(modifier = Modifier.height(screenHeight * 0.02f))
-
-                    HorizontalDivider(
-                        modifier =
-                            Modifier.padding(horizontal = screenWidth * 0.04f)
-                                .testTag("sliding_window_horizontal_divider_3"),
-                        thickness = 1.dp,
-                        color = colorScheme.onSurface.copy(alpha = 0.2f))
-                    Spacer(modifier = Modifier.height(screenHeight * 0.02f))
-
-                    Text(
-                        text = "Reviews",
-                        style = MaterialTheme.typography.headlineMedium,
-                        color = colorScheme.onBackground,
-                        modifier = Modifier.padding(horizontal = screenWidth * 0.04f))
-                    Spacer(modifier = Modifier.height(screenHeight * 0.01f))
-
-                    // Star Rating Row
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically,
-                        modifier =
-                            Modifier.padding(horizontal = screenWidth * 0.04f)
-                                .testTag("sliding_window_star_rating_row")) {
-                          val filledStars = workerRating.toInt()
-                          val unfilledStars = 5 - filledStars
-                          repeat(filledStars) {
-                            Icon(
-                                imageVector = Icons.Filled.Star,
-                                contentDescription = null,
-                                tint = colorScheme.onBackground)
-                          }
-                          repeat(unfilledStars) {
-                            Icon(
-                                imageVector = Icons.Outlined.StarOutline,
-                                contentDescription = null,
-                                tint = colorScheme.onBackground,
-                            )
-                          }
-                        }
-                    Spacer(modifier = Modifier.height(screenHeight * 0.01f))
-                    LazyRow(
-                        modifier =
-                            Modifier.fillMaxWidth()
-                                .padding(horizontal = screenWidth * 0.04f)
-                                .testTag("sliding_window_reviews_row")) {
-                          itemsIndexed(reviews) { index, review ->
-                            var isExpanded by remember { mutableStateOf(false) }
-                            val displayText =
-                                if (isExpanded || review.length <= 100) {
-                                  review
-                                } else {
-                                  review.take(100) + "..."
-                                }
-
-                            Box(
-                                modifier =
-                                    Modifier.padding(end = screenWidth * 0.02f)
-                                        .width(screenWidth * 0.6f)
-                                        .clip(RoundedCornerShape(25f))
-                                        .background(colorScheme.background)) {
-                                  Column(modifier = Modifier.padding(screenWidth * 0.02f)) {
-                                    Text(
-                                        text = displayText,
-                                        style = MaterialTheme.typography.bodySmall,
-                                        color = colorScheme.onSurface)
-                                    if (review.length > 100) {
-                                      Text(
-                                          text = if (isExpanded) "See less" else "See more",
-                                          style =
-                                              MaterialTheme.typography.bodySmall.copy(
-                                                  color = colorScheme.primary),
-                                          modifier =
-                                              Modifier.clickable { isExpanded = !isExpanded }
-                                                  .padding(top = screenHeight * 0.01f))
-                                    }
-                                  }
-                                }
-                          }
-                        }
-
-                    Spacer(modifier = Modifier.height(screenHeight * 0.02f))
-                  }
-            }
-      }
+    // Call to QuickFixSlidingWindowWorker
+    QuickFixSlidingWindowWorker(
+        isVisible = isWindowVisible,
+        onDismiss = { isWindowVisible = false },
+        bannerImage = bannerImage,
+        profilePicture = profilePicture,
+        initialSaved = saved,
+        workerCategory = workerCategory,
+        workerAddress = workerAddress,
+        description = description,
+        includedServices = includedServices,
+        addonServices = addonServices,
+        workerRating = workerRating,
+        tags = tags,
+        reviews = reviews,
+        screenHeight = screenHeight,
+        screenWidth = screenWidth,
+        onContinueClick = { /* Handle continue */})
   }
 }
